refactor: remove backwards-compat aliases and inbox-types wrapper
- Remove 8 unused getWorkspace*/fetchWorkspace* function aliases from
  settings-utils.ts (zero consumers, pure migration shims)
- Rename useWorkspaceLogo→useSettingsLogo, useWorkspaceHeaderLogo→
  useSettingsHeaderLogo for naming consistency
- Delete inbox-types.ts re-export wrapper, update 5 consumers to import
  directly from @/lib/shared/types
- Add doc comment to getRedisConnectionOpts explaining why it returns
  a new object per call (BullMQ requirement)
