+++ apps/web/src/components/admin/feedback/inbox-container.tsx
@@ -10,7 +10,7 @@ import { useInboxFilters } from '@/components/admin/feedback/use-inbox-filters'
 import { useInboxPosts, flattenInboxPosts, inboxKeys } from '@/lib/client/hooks/use-inbox-query'
 import { useSegments } from '@/lib/client/hooks/use-segments-queries'
 import { mergeSuggestionQueries } from '@/lib/client/queries/signals'
-import type { CurrentUser } from '@/components/admin/feedback/inbox-types'
+import type { CurrentUser } from '@/lib/shared/types'
 import type { Board, Tag, InboxPostListResult, PostStatusEntity } from '@/lib/shared/db-types'
 import type { TeamMember } from '@/lib/server/domains/principals'
 import type { PostId } from '@quackback/ids'
