--- apps/web/src/components/admin/feedback/inbox-types.ts
@@ -1,12 +1,0 @@
-/**
- * Re-exports from @/lib/types/inbox for backwards compatibility.
- *
- * Prefer importing directly from '@/lib/shared/types' in new code.
- */
-export type {
-  PinnedComment,
-  CommentReaction,
-  CommentWithReplies,
-  PostDetails,
-  CurrentUser,
-} from '@/lib/shared/types'
