+++ apps/web/src/components/admin/feedback/suggestions/merge-preview-modal.tsx
@@ -24,7 +24,7 @@ import {
   getInitialContentJson,
 } from '@/components/admin/feedback/detail/post-utils'
 import type { PostId, CommentId } from '@quackback/ids'
-import type { PostDetails } from '@/components/admin/feedback/inbox-types'
+import type { PostDetails } from '@/lib/shared/types'
 import type { PublicCommentView } from '@/lib/client/queries/portal-detail'
 
 interface MergePreviewModalProps {
