+++ apps/web/src/lib/client/hooks/use-settings-queries.ts
@@ -1,7 +1,7 @@
 /**
  * Settings query hooks
  *
- * Query hooks for workspace settings.
+ * Query hooks for settings data.
  * Mutations are in @/lib/client/mutations/settings.
  */
 
@@ -12,14 +12,14 @@ import { settingsQueries } from '@/lib/client/queries/settings'
 // Query Hooks
 // ============================================================================
 
-export function useWorkspaceLogo() {
+export function useSettingsLogo() {
   return useQuery({
     ...settingsQueries.logo(),
     enabled: false, // Use SSR data, don't auto-fetch
   })
 }
 
-export function useWorkspaceHeaderLogo() {
+export function useSettingsHeaderLogo() {
   return useQuery({
     ...settingsQueries.headerLogo(),
     enabled: false, // Use SSR data, don't auto-fetch
